+++ src/main/java/dev/salonce/discordquizbot/application/RawQuestionsService.java
@@ -6,6 +6,7 @@ import dev.salonce.discordquizbot.domain.DifficultyLevel;
 import dev.salonce.discordquizbot.infrastructure.storage.RawQuestionStore;
 import dev.salonce.discordquizbot.infrastructure.configs.TopicsConfig;
 import dev.salonce.discordquizbot.infrastructure.dtos.RawQuestion;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -14,41 +15,12 @@ import java.util.stream.Collectors;
 
 @Slf4j
 @Service
+@RequiredArgsConstructor
 public class RawQuestionsService {
 
     private final RawQuestionStore rawQuestionStore;
-    private final Categories categories = new Categories();
 
-    public RawQuestionsService(TopicsConfig topicsConfig, RawQuestionStore rawQuestionStore){
-        this.rawQuestionStore = rawQuestionStore;
-
-        for (Map.Entry<String, Set<String>> entry : topicsConfig.getAvailableTopics().entrySet()) {
-            String topicName = entry.getKey();
-            Set<String> tagsSet = entry.getValue();
-            List<RawQuestion> rawTopicQuestions = getRawQuestionsForTags(tagsSet);
-            List<DifficultyLevel> difficultyLevels = prepareDifficultyLevels(rawTopicQuestions);
-            Category category = new Category(topicName, difficultyLevels);
-            categories.addTopic(topicName, category);
-        }
-    }
-
-    public Categories getTopics() {
-        return categories;
-    }
-
-    public boolean areNoTopicsAvailable(){
-        return categories.areNone();
-    }
-
-    public boolean doesQuestionSetExist(String topic, int level){
-        return categories.doesQuestionSetExist(topic, level);
-    }
-
-    public List<RawQuestion> getRawQuestionList(String topic, int level){
-        return categories.getRawQuestionList(topic, level);
-    }
-
-    public List<RawQuestion> removePrepareQuestionsForDifficultyLevel(List<RawQuestion> removableRawQuestions) {
+    private List<RawQuestion> removePrepareQuestionsForDifficultyLevel(List<RawQuestion> removableRawQuestions) {
         List<RawQuestion> preparedRawQuestions = new ArrayList<>();
         if (removableRawQuestions.size() < 65) {
             int size = removableRawQuestions.size();
@@ -65,7 +37,7 @@ public class RawQuestionsService {
         return preparedRawQuestions;
     }
 
-    private List<DifficultyLevel> prepareDifficultyLevels(List<RawQuestion> rawQuestions) {
+    public List<DifficultyLevel> prepareDifficultyLevels(List<RawQuestion> rawQuestions) {
         rawQuestions.sort(Comparator
                 .comparing(RawQuestion::difficulty, Comparator.nullsLast(Integer::compareTo))
                 .thenComparing(RawQuestion::id, Comparator.nullsLast(Long::compareTo)));
@@ -80,7 +52,7 @@ public class RawQuestionsService {
         return difficulties;
     }
 
-    private List<RawQuestion> getRawQuestionsForTags(Set<String> topicTags) {
+    public List<RawQuestion> getRawQuestionsForTags(Set<String> topicTags) {
         return rawQuestionStore.getRawQuestions().stream()
                 .filter(rawQuestion -> {
                     Set<String> rawQuestionTags = rawQuestion.tags();
